#180417061 NPE fix (#2)
Ticket: https://www.pivotaltracker.com/story/show/180417061

What changed?
======================
We’re now returning the pagination URL even if there are errors present in the table list response. 

On the E2E test side, we’re no longer asserting that the response has no errors. This is now logged as a warning instead.

Why are we doing this?
=====================
To ensure that our E2E test is able to iterate through all the catalogs shown in the index array of the /tables call even if some catalogs have errors.