[#179810521] Cleanup old queries (#3)
Ticket: https://www.pivotaltracker.com/story/show/179810521

What changed?
======================
We’ve implemented a @Scheduled method that runs every 10s, to cleanup queries in Trino that haven’t finished and haven’t accessed in the last 2 minutes.

Why are we doing this?
=====================
It came up that our micro-services that call data-connect-trino/trino don't cancel the query when they shut down, so there are a lot of queries that are “stuck” in trino till trino is restarted. We want to cleanup these queries to reduce load on Trino.