[#181754324] Fix auditing (#10)
Ticket: https://www.pivotaltracker.com/story/show/181754324

What changed?
======================
We fixed the auditing in the data-connect-trino so that we’re not manually logging events but replying on the auto logging from the spring-boot-audit-event-logger lib. We also added custom audit action uri based on the endpoint,  masked the users token in the audit events, and made minor code improvements.

Why are we doing this?
=====================
Because we need better auditing for data-connect-trino, and they should be in line with our other audit events.