[#186583773] - updated query to get tables in trino
Why:
the old query stopped returning views in the results. For some reason, making the qeury more specific returns the missing view results.
This seems like a bug with trino, but we haven't been able to reproduce it in a native trino connector
so maybe there's another bug with our code that we haven't discovered yet.

Co-authored-by: Jonathan <jonathan@dnastack.com>