[CU-86b17aw6h] Add support for faking https to trino
In order to use an Authenticator, Trino needs to think it is running HTTPS
or that HTTPS is being terminated for it by a proxy.

These changes add the X-Forwarded-Proto to the request, and rewrite the
schema of the next page uri in the response so as to fool trino into
thinking it is behind an SSL terminating proxy.