[CU-86b6f3a36] Fix query timeouts and slow starts
We determined through experimentation that Trino does not expect clients to sleep between next page requests. Trino holds on to next page requests for up to 10 seconds when there is no more data available. The client should call back immediately for the next page as soon as it receives a response.

This change to the retry interval has made small queries go much faster. We will monitor this change for potential client timeouts and its effects on 401 errors due to token expiration.

This commit also removes the default "scope" auth mechanism, which was not working and had no test coverage. If someone needs that kind of auth in the future, they should develop a new bean for that and include test coverage.

Finally, we removed the direct calls to Trino from the e2e suite, preferring a symmetric data path where we write and read all data through the system under test.

Co-authored-by: Kevin Hong <kevin@dnastack.com>